NLP/UniMS/requirements/requirements.txt
detectron2==0.1
ftfy==6.0.3
h5py==2.7.1
matplotlib==2.1.2
nltk==3.2.5
numpy==1.19.5
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==6.2.2
prefetch-generator==1.0.1
pytorch-lightning==1.5.10
rouge==1.0.0
scikit-learn==0.19.1
scipy==1.5.1
spacy==3.2.3
timm==0.5.4
torch==1.8.1
torchmetrics==0.7.2
torchvision==0.11.2
tqdm==4.63.0
transformers==4.17.0
wandb==0.12.10